Menambahkan modifikasi untuk PowerShell dan WindowsPowerShell Configurations

+++ Main-PC-Laptop/Windows-dotfiles/WindowsPowershell/Microsoft.PowerShell_profile.ps1
@@ -1,24 +1,24 @@
 
-# Konfigurasi PSReadLine untuk PowerShell
+# Konfigurasi PSReadLine untuk Windows PowerShell 5.1
 # yang Menyediakan dukungan untuk line editing dan keybinding.
 
 # Import PSReadLine
 Import-Module PSReadLine
 
 # Keybinding untuk Readline Emacs
-Set-PSReadLineKeyHandler -Key Ctrl+p -Function PreviousHistory
-Set-PSReadLineKeyHandler -Key Ctrl+n -Function NextHistory
-Set-PSReadLineKeyHandler -Key Ctrl+f -Function ForwardChar
-Set-PSReadLineKeyHandler -Key Ctrl+b -Function BackwardChar
-Set-PSReadLineKeyHandler -Key Ctrl+a -Function BeginningOfLine
-Set-PSReadLineKeyHandler -Key Ctrl+e -Function EndOfLine
+Set-PSReadLineKeyHandler -Chord Ctrl+p -Function PreviousHistory
+Set-PSReadLineKeyHandler -Chord Ctrl+n -Function NextHistory
+Set-PSReadLineKeyHandler -Chord Ctrl+f -Function ForwardChar
+Set-PSReadLineKeyHandler -Chord Ctrl+b -Function BackwardChar
+Set-PSReadLineKeyHandler -Chord Ctrl+a -Function BeginningOfLine
+Set-PSReadLineKeyHandler -Chord Ctrl+e -Function EndOfLine
 
 # Keybinding untuk maju dan mundur satu kata
-Set-PSReadLineKeyHandler -Key Alt+f -Function ForwardWord
-Set-PSReadLineKeyHandler -Key Alt+b -Function BackwardWord
+Set-PSReadLineKeyHandler -Chord Alt+f -Function ForwardWord
+Set-PSReadLineKeyHandler -Chord Alt+b -Function BackwardWord
 
 # Keybinding untuk Undo
-Set-PSReadLineKeyHandler -Key Ctrl+/ -Function Undo
+Set-PSReadLineKeyHandler -Chord Ctrl+/ -Function Undo
 
 # Custom Script Folder
 $env:PATH += ";C:\Users\acer8\Documents\WindowsPowerShell\Scripts"
@@ -28,7 +28,9 @@ Set-Alias -Name btop -Value "btop4win.exe"
 Set-Alias -Name vim -Value nvim
 
 # Remove alias default bawaan
-Remove-Item Alias:ls
+if (Test-Path Alias:ls) {
+    Remove-Item Alias:ls
+}
 
 # Function PowerShell Tambahan (WIP)
 function yt-mp3 {
@@ -36,13 +38,14 @@ function yt-mp3 {
 }
 
 function tp-cmd-alias {
-	cd "C:\Users\acer8\Documents\CMD\aliases"
+    cd "C:\Users\acer8\Documents\CMD\aliases"
 }
 
 function ls {
-	lsd -la1hF $args
+    lsd -la1hF $args
 }
 
 # Menjalankan fastfetch saat PowerShell dibuka
 fastfetch
 Invoke-Expression (&scoop-search --hook)
+
